+++ ios/Tests/ByteowlsCapacitorFilesharerTests/ByteowlsCapacitorFilesharerTests.swift
@@ -1,0 +1,8 @@
+import XCTest
+@testable import ByteowlsCapacitorFilesharer
+
+final class ByteowlsCapacitorFilesharerTests: XCTestCase {
+    func testPluginCanBeCreated() {
+        XCTAssertNotNil(FileSharerPlugin())
+    }
+}